\name{print.threg}
\alias{print.threg}
\title{Print method for threg objects}
\description{
 Produces a printed summary of a fitted threg model 
 }
\usage{
\method{print}{threg}(x, digits=max(options()$digits - 4, 3), ...)
}

\arguments{
  \item{x}{
    the result of a call to \code{threg} 
  }
  \item{digits}{significant digits to print}
  \item{\dots}{For future methods}
}

